Add logging and clean up implemetnation

+++ app/controller/todo-list-controller.js
@@ -1,19 +1,19 @@
 'use strict';
 
-const _ = require('lodash');
 const Joi = require('joi');
 const {InternalServerError, BadRequestError, NotFoundError} = require('restify-errors');
-const {listValidatorSchema, listIdValdatorSchema} = require('./validation');
+const {listValidator, listIdValdator} = require('./validation');
 
-const todoListDao = require('../db/todo-db-list');
+const {todoListDao} = require('../db/todo-db-list');
+const {todoEntryDao} = require('../db/todo-db-entry');
 const errors = require('../model/errors');
-const logger = require('../util/logger');
+const {logger} = require('../util/logger');
 
 
 async function createList(request, response, next) {
 
   let input = request.body;
-  let validationResult = Joi.validate(input, listValidatorSchema);
+  let validationResult = Joi.validate(input, listValidator);
 
   if (validationResult.error) {
     logger.debug('createList - validation error: %j', validationResult.error);
@@ -23,11 +23,11 @@ async function createList(request, response, next) {
   let listName = input.name;
   logger.debug('createList - process request: %j', input);
   try {
-    let list = await todoListDao.createList(listName);
+    let {list} = await todoListDao.createList(listName);
     response.status(201);
     return handleSuccess(list, response, next, 'createList - New list created successfully');
   } catch (error) {
-    logger.error('Error trying to create list: %s', error);
+    logger.error('createList - Error trying to create list: %s', error);
     return next(new InternalServerError(`Error trying to create list ${listName}`));
   }
 
@@ -36,7 +36,7 @@ async function createList(request, response, next) {
 async function getList(request, response, next) {
 
   let listId = request.params.listId;
-  let validationResult = Joi.validate(listId, listIdValdatorSchema);
+  let validationResult = Joi.validate(listId, listIdValdator);
 
   if (validationResult.error) {
     let error = validationResult.error;
@@ -47,14 +47,15 @@ async function getList(request, response, next) {
   logger.debug('getList - for list with id: %d', listId);
   try {
     let list = await todoListDao.findById(listId);
-    if (list) {
-      return handleSuccess(list, response, next, 'get - List found');
-    } else {
-      return next(new NotFoundError(errors.ListNotFound));
-    }
+
+    if (!list) {
+      logger.debug('getList - No list found with id: %d', listId);
+      return next(new NotFoundError(errors.ListNotFound.toObject()));
+    } 
     
+    return handleSuccess(list, response, next, 'get - List found');
   } catch (error) {
-    logger.error('Error trying to retrieve list: %s', error);
+    logger.error('getList - Error trying to retrieve list: %s', error);
     return next(new InternalServerError(`Error trying to create list with id ${listId}`));
   }
 
@@ -62,16 +63,25 @@ async function getList(request, response, next) {
 
 async function deleteList(request, response, next) {
   let listId = request.params.listId;
-  let validationResult = Joi.validate(listId, listIdValdatorSchema);
+  let validationResult = Joi.validate(listId, listIdValdator);
 
   if (validationResult.error) {
     let error = validationResult.error;
-    logger.debug('deleteList - validation error: %j', error);
+    logger.debug('deleteList - list id validation error: %j', error);
     return next(createInvalidListIdError(error));
   } 
 
   logger.debug('deleteList - trying to delete list with id: %d', listId);
   try {
+
+    let list = await todoListDao.findById(listId);
+    
+    if (!list) {
+      logger.debug('deleteList - No list found with id: %d', listId);
+      return next(new NotFoundError(errors.ListNotFound.toObject()));
+    } 
+
+    await todoEntryDao.deleteListTodos(list);
     await todoListDao.deleteById(listId);
     logger.debug('deleteList - list deleted with id: %d', listId);
     response.status(204);
@@ -95,24 +105,23 @@ function handleSuccess(list, response, next, message) {
 }
 
 function handleCreateValidationError(error, next) {
-  let additionalInfo = { additionalInfo: error.details.message };
   let errorResponse;
 
   if (error.details[0] && error.details[0].type === 'any.unknown') {
-    errorResponse = _.assign({}, errors.IdNotAcceptable, additionalInfo);
+    errorResponse = errors.IdNotAcceptable;
   } else {
-    errorResponse = _.assign({}, errors.InvalidListName, additionalInfo);
+    errorResponse = errors.InvalidListName;
   }
-  return next(new BadRequestError(errorResponse));
+  errorResponse = errorResponse.withAdditionalInfo(error.details.message);
+  return next(new BadRequestError(errorResponse.toObject()));
 }
 
 function createInvalidListIdError(error) {
-  let additionalInfo = { additionalInfo: error.details.message };
-  let  errorResponse = _.assign({}, errors.IdNotAcceptable, additionalInfo);
-  return new BadRequestError(errorResponse);
+  let errorResponse = errors.IdNotAcceptable.withAdditionalInfo(error.details.message);
+  return new BadRequestError(errorResponse.toObject());
 }
 
-class TodoListConstroller {}
+function TodoListConstroller () {}
 TodoListConstroller.prototype.createList = createList;
 TodoListConstroller.prototype.getList = getList;
 TodoListConstroller.prototype.deleteList = deleteList;
